refactor: use flat structure in common/cards folder
