+++ app/(pages)/visit/page.tsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import { CardGroup, Grid, GridContainer } from '@trussworks/react-uswds';
 import { Paragraph } from 'app/components/common/Paragraph';
-import ExhibitCard from 'app/components/common/cards/small/exhibit';
+import ExhibitCard from 'app/components/common/cards/ExhibitCard';
 import { DATA_EXHIBITS as exhibits } from 'app/constants';
 
 const VisitPage: React.FC = () => {
