+++ app/components/common/cards/ExhibitCard.tsx
@@ -1,7 +1,6 @@
-'use client';
 import React from 'react';
 import { Icon } from '@trussworks/react-uswds';
-import SmallCard from '../index';
+import SmallCard from './SmallCard';
 import type { Exhibit } from 'app/types';
 
 interface FormattedSectionProps {
